pyproject.toml
[project]
name = "napari-timelapse-capture"
version = "2.0.0"  # ✅ Refactored version with modular architecture
description = "Modular Napari plugin for Nematostella timelapse recording with ESP32 LED synchronization and HDF5 data management"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "s1alknau", email = "akknauss@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: napari",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "napari>=0.4.15",
    "magicgui",
    "qtpy>=2.0",
    "h5py>=3.0",          # For HDF5 data storage
    "pyserial>=3.5",      # For ESP32 communication
    "Pillow>=9.0",        # For image I/O (TIFF, PNG)
]

[project.optional-dependencies]
# Testing dependencies
testing = [
    "tox",
    "pytest>=7.0",
    "pytest-cov",
    "pytest-qt",
    "pytest-mock",
    "napari",
    "pyqt5",
]

# Development dependencies
dev = [
    "black",              # Code formatting
    "ruff",               # Linting
    "mypy",               # Type checking
    "pre-commit",         # Git hooks
]

# Documentation dependencies
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
]

# Optional: OpenCV for advanced image processing
opencv = [
    "opencv-python>=4.5",
]

# Optional: IMSwitch integration
imswitch = [
    "imswitch",
]

# All optional dependencies
all = [
    "opencv-python>=4.5",
    "imswitch",
]

[project.urls]
Homepage = "https://github.com/s1alknau/napari-timelapse-capture"
"Bug Tracker" = "https://github.com/s1alknau/napari-timelapse-capture/issues"
Documentation = "https://github.com/s1alknau/napari-timelapse-capture#readme"
"Source Code" = "https://github.com/s1alknau/napari-timelapse-capture"

[project.entry-points."napari.manifest"]
napari-timelapse-capture = "timeseries_capture:napari.yaml"

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["timeseries_capture*"]
exclude = ["tests*", "docs*", "*.egg-info", "Backup*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

# ============================================================================
# BLACK - Code Formatting
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | Backup
  | Backup_2
  | Backup3
)/
'''

# ============================================================================
# RUFF - Fast Python Linter
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".tox",
    ".venv",
    "build",
    "dist",
    "*.egg-info",
    "Backup",
    "Backup_2",
    "Backup3",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# ============================================================================
# MYPY - Type Checking
# ============================================================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter checking
ignore_missing_imports = true
exclude = [
    "Backup",
    "Backup_2",
    "Backup3",
    "tests",
]

# ============================================================================
# PYTEST - Testing
# ============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=timeseries_capture",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "hardware: marks tests that require hardware (ESP32, camera)",
    "gui: marks tests that require GUI",
]

# ============================================================================
# COVERAGE - Code Coverage
# ============================================================================
[tool.coverage.run]
source = ["timeseries_capture"]
omit = [
    "*/tests/*",
    "*/Backup*/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
